+++ src/cast.rs
@@ -15,7 +15,7 @@
 
 use crate::{
     FixedI128, FixedI16, FixedI32, FixedI64, FixedI8, FixedU128, FixedU16, FixedU32, FixedU64,
-    FixedU8, F128,
+    FixedU8,
 };
 use az::{Cast, CheckedCast, OverflowingCast, SaturatingCast, UnwrappedCast, WrappingCast};
 use half::{bf16 as half_bf16, f16 as half_f16};
@@ -186,7 +186,7 @@ macro_rules! cast_num {
             $Fixed($nbits);
             i8, i16, i32, i64, i128, isize,
             u8, u16, u32, u64, u128, usize,
-            half_f16, half_bf16, f32, f64, F128,
+            half_f16, half_bf16, f32, f64, f128,
         }
     )* };
 }